backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Validation and settings
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Async HTTP client
httpx>=0.26.0
aiofiles>=23.2.1

# Task queue
celery[redis]>=5.3.6
redis>=5.0.1
flower>=2.0.1

# AI/ML
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.1

# Audio processing
scipy>=1.11.4
numpy>=1.26.3
soundfile>=0.12.1
pydub>=0.25.1

# Claude API
anthropic>=0.39.0

# Content extraction
beautifulsoup4>=4.12.2
readability-lxml>=0.8.1
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.3
python-slugify>=8.0.1

# Rate limiting
slowapi>=0.1.9

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
ruff>=0.1.11
